+++ lib/oli/authoring/course/project.ex
@@ -10,6 +10,7 @@ defmodule Oli.Authoring.Course.Project do
     field :slug, :string
     field :title, :string
     field :version, :string
+    field :visibility, Ecto.Enum, values: [:authors, :selected, :global], default: :authors
 
     belongs_to :parent_project, Oli.Authoring.Course.Project, foreign_key: :project_id
     belongs_to :family, Oli.Authoring.Course.Family
@@ -23,7 +24,7 @@ defmodule Oli.Authoring.Course.Project do
   @doc false
   def changeset(project, attrs \\ %{}) do
     project
-      |> cast(attrs, [:title, :slug, :description, :version, :family_id, :project_id])
+      |> cast(attrs, [:title, :slug, :description, :version, :family_id, :project_id, :visibility])
       |> validate_required([:title, :version, :family_id])
       |> Slug.update_never("projects")
   end